Add CI for matching service and Refactor handler to controller
