+++ matching-service/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "nodemon index.js",
     "start": "npm run db:migrate:production && node index.js",
+    "test:ci": "npm run db:migrate:production && npm run lint && mocha ./test/*.spec.js --exit",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint .",
     "lint:fix": "eslint --fix --ext .js .",
@@ -18,6 +19,8 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "axios-mock-adapter": "^1.21.2",
+    "chai": "^4.3.6",
     "eslint": "^8.2.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-node": "^4.1.0",
@@ -25,9 +28,11 @@
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "mocha": "^10.1.0",
     "nodemon": "^2.0.19",
     "prettier": "^2.7.1",
-    "sequelize-cli": "^6.4.1"
+    "sequelize-cli": "^6.4.1",
+    "supertest": "^6.3.1"
   },
   "dependencies": {
     "axios": "^0.27.2",
