+++ room-service/.env.prod
@@ -1,0 +1,6 @@
+ACCESS_TOKEN_SECRET=7c0dd92cafc2537595f48eca543a45e38e25b0322c558b8cb44907db49726b54a0809709c4491cbc086187009367ac4febfb8d5db310320c16d510818d833ad9
+POSTGRES_HOST='host.docker.internal'
+POSTGRES_USER='postgres'
+POSTGRES_DB='postgres_db'
+POSTGRES_PASSWORD='postgres'
+POSTGRES_PORT=5432
